pyproject.toml
[project]
name = "nonebot-plugin-ottohzys"
dynamic = ["version"]
description = "otto!"
authors = [{ name = "LgCookie", email = "lgc2333@126.com" }]
dependencies = [
    "nonebot2>=2.4.2",
    "nonebot-plugin-alconna>=0.56.2",
    "nonebot-plugin-localstore>=0.7.4",
    "cookit[pydantic,nonebot-localstore]>=0.10.2",
    "soundfile>=0.13.1",
    "numpy>=2.2.4",
    "pypinyin>=0.54.0",
    "psola>=0.0.1",
    "anyio>=4.9.0",
    "httpx>=0.28.1",
    "async-lru>=2.0.5",
]
requires-python = ">=3.9,<4.0"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
homepage = "https://github.com/lgc-NB2Dev/nonebot-plugin-ottohzys"

[tool.pdm.build]
includes = []

[tool.pdm.version]
source = "file"
path = "nonebot_plugin_ottohzys/__init__.py"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
